requirements.txt
streamlit
pandas
black
flake8
pytest
python-dotenv
